[TASK] Configure git during build phase as well
Git config (e.g. user name, email) was not set during `docker build`
phase. Because we configure it during `docker run` phase, it's not a big
deal to include it here too.

This might be especially useful when calling hook_user_build_script,
which might do some git commands which might require it to be configured.